Add message that SRV check was skipped if .well-known is configured
The backend does not check for the SRV records if a .well-known
configuration with a port declaration was found. It will return an
empty SRVRecords field which leads to the understanding that nothing
was found even though it was intentionally skipped. The newly introduced
message will inform the user that it was skipped because a .well-known
configuration was found.

Refs:
- https://github.com/matrix-org/matrix-federation-tester/issues/104
- https://github.com/matrix-org/matrix-federation-tester/pull/116