Always include Hosts informations inside the dns result section
Until now only the address information was displayed if some dns records
where found. It will now not only display the full information when the
dns records are present, but also when the lookup was skiped due to the
presents of a .wellknown file with a explicite port number.

Whith this changes it will not longer check if there are one or more
hosts. I honestly don't know if this will create an error in some
specific cases.